chore(metrics): share metric summary math as a reusable module
Extract the pure total/average/latest summary + change-pill math out of the
insights scene into lib/components/Metric/metricSummary so the Metrics product
can reuse it without importing from scenes/insights. The insights Metric.utils
module now re-exports from the shared location, so existing call sites and the
unit test are unchanged.

Generated-By: PostHog Code
Task-Id: aa86d15f-4155-4e29-a2dd-1027669fa8ee
